SWT 3.3.1.1 version bump (fixes Eclipse printing problems).
git-svn-id: https://overlays.gentoo.org/svn/proj/java/java-experimental@5784 fe387f86-7aec-0310-8f44-a148a97195aa

+++ eclass/osgi.eclass
@@ -199,6 +199,8 @@ _java-pkg_osgijar-fromfile() {
 
 	# We automatically change the version unless --noversion is present
 
+	[[ -e "${arguments[1]}" ]] || die "Manifest file ${arguments[1]} not found"
+				
 	if [[ "${noversion}" == 1 ]]; then
 		cat "${arguments[1]}" > "${_OSGI_T}/tmp_jar/META-INF/MANIFEST.MF"
 	else
